pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-ads-mcp"
version = "0.1.0"
description = "Google Ads & GA4 MCP server for Claude Code and Claude Desktop"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Gabriel Urrutia", email = "" }
]
keywords = ["google-ads", "ga4", "mcp", "claude", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=2.0.0",
    "google-ads>=29.0.0",
    "google-analytics-data>=0.18.0",
    "google-analytics-admin>=0.22.0",
    "google-auth-oauthlib>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
google-ads-mcp = "google_ads_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/gabogabucho/google-ads-mcp"
"Bug Tracker" = "https://github.com/gabogabucho/google-ads-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/google_ads_mcp"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
